[ftr tests] split x-pack functional_with_es_ssl config
