+++ x-pack/test/functional_with_es_ssl/config.base.ts
@@ -6,7 +6,7 @@
  */
 
 import Fs from 'fs';
-import { resolve, join } from 'path';
+import { join } from 'path';
 import { CA_CERT_PATH } from '@kbn/dev-utils';
 import { FtrConfigProviderContext } from '@kbn/test';
 import { pageObjects } from './page_objects';
@@ -51,14 +51,7 @@ export default async function ({ readConfigFile }: FtrConfigProviderContext) {
     ...xpackFunctionalConfig.getAll(),
     servers,
     pageObjects,
-    // list paths to the files that contain your plugins tests
-    testFiles: [
-      resolve(__dirname, './apps/triggers_actions_ui'),
-      resolve(__dirname, './apps/discover'),
-      resolve(__dirname, './apps/uptime'),
-      resolve(__dirname, './apps/ml'),
-      resolve(__dirname, './apps/cases'),
-    ],
+    // Don't list paths to the files that contain your plugins tests here
     apps: {
       ...xpackFunctionalConfig.get('apps'),
       triggersActions: {
